Clearer naming of array based ephemeral caches
